pyproject.toml
[project]
name = "crypto-ai-service"
version = "0.1.0"
description = "MacBook MPS + PyTorch 기반 암호화폐 AI 분석 서비스"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "개발자", email = "dev@example.com" }
]
keywords = ["cryptocurrency", "ai", "pytorch", "analysis", "mps"]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "python-binance>=1.0.0",  # Binance API wrapper
]

[project.optional-dependencies]
api = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
]
viz = [
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "tensorboard>=2.14.0",
]
all = [
    "crypto-ai-service[api,viz,dev]",
]

[project.scripts]
crypto-ai = "crypto_ai.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crypto_ai"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
